Removed pfglue/prmsieve as they are obsolete.  Replaced primegen with new sieve code in primeserver.cpp
git-svn-id: https://svn.code.sf.net/p/openpfgw/code@508 368e28a5-49e5-40c5-be6c-d84989c25d7b

+++ pform/pfgwlib/gwcontext.cpp
@@ -34,7 +34,7 @@ int  StripTrailingWhiteSpace(char *string)
 {
   long  x;
 
-  x = strlen(string);
+  x = (long) strlen(string);
 
   while (x > 0 && (string[x-1] == ' '))
   {
